requirements.txt
langchain
langgraph
langchain-openai
python-dotenv
redis
chromadb
tiktoken
pydantic
google-genai
dotenv